// src/modules/emitir/dto/otra.moneda.dto.ts
export class OtraMoneda {
    TipoMoneda: string;
    TipoCambio: string;
    TotalOtraMoneda: string;



    xml: string = `<OtraMoneda> 
                            <TipoMoneda>CLP</TipoMoneda> 
                            <TipoCambio> 0</TipoCambio> 
                            <TotalOtraMoneda> 0.00</TotalOtraMoneda> 
                           </OtraMoneda>`;
}